requirements.txt
# MCP SDK
mcp>=1.0.0

# Claude API
anthropic>=0.40.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# HTTP
requests>=2.31.0

# Environment
python-dotenv>=1.0.0
